+++ SMS.Contracts/Requests/ApplicationLogs/ApplicationLogRequestDto.cs
@@ -2,7 +2,7 @@
 {
     public class ApplicationLogRequestDto
     {
-        public int? AuditLogId { get; set; }
+        public long? AuditLogId { get; set; }
         public required string Message { get; set; }
         public Exception? Exception { get; set; }
         public string? StackTrace { get; set; }
